Remove party.party.default.bank_account model.
Move back to previous way of managing bank accounts. Current design does not
support a way to define which company's account is going to be used and anyway
using a MatchMixin seems excessive for such a narrow use case. We probably need
to add bank account and payment type to models that don't have them and in
doing so, the need for a MatchMixin vanishes.